some cleaning

+++ beatrun/gamemodes/beatrun/gamemode/player_class/player_beatrun.lua
@@ -357,8 +357,9 @@ end
 
 function PLAYER:CalcView(view)
 	if CLIENT then
-		view.fov = GetConVar("Beatrun_FOV"):GetFloat()
+		view.fov = GetConVar("Beatrun_FOV"):GetInt()
 	end
+
 	if self.TauntCam:CalcView(view, self.Player, self.Player:IsPlayingTaunt()) then return true end
 end
